+++ albertdesk/__init__.py
@@ -1,10 +1,10 @@
 """
 AlbertDesk - Remote Desktop Control Software
-Version: 1.1.0
+Version: 1.2.0
 A Professional RustDesk-like application for remote desktop sharing and control.
 """
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 __author__ = "Albert"
 __license__ = "MIT"